+++ src/listener/formatter/builder.js
@@ -1,23 +1,44 @@
+import _ from 'lodash'
+import getColorFns from '../../get_color_fns'
+import JavascriptSnippetSyntax from '../../step_definition_snippet_builder/javascript_snippet_syntax'
 import JsonFormatter from './json'
 import PrettyFormatter from './pretty'
 import ProgressFormatter from './progress'
 import RerunFormatter from './rerun'
+import StepDefinitionSnippetBuilder from '../../step_definition_snippet_builder'
 import SummaryFormatter from './summary'
 
-function getConstructorByType(type) {
-  switch(type) {
-    case 'json': return JsonFormatter
-    case 'pretty': return PrettyFormatter
-    case 'progress': return ProgressFormatter
-    case 'rerun': return RerunFormatter
-    case 'summary': return SummaryFormatter
-    default: throw new Error('Unknown formatter name "' + type + '".')
+export default class FormatterBuilder {
+  static build(type, options) {
+    const Formatter = FormatterBuilder.getConstructorByType(type)
+    const extendedOptions = _.assign({}, options, {
+      colorFns: getColorFns(options.colorsEnabled),
+      snippetBuilder: FormatterBuilder.getStepDefinitionSnippetBuilder(options)
+    })
+    return new Formatter(extendedOptions)
   }
-}
 
-export default {
-  build(type, options) {
-    const Formatter = getConstructorByType(type)
-    return new Formatter(options)
+  static getConstructorByType(type) {
+    switch(type) {
+      case 'json': return JsonFormatter
+      case 'pretty': return PrettyFormatter
+      case 'progress': return ProgressFormatter
+      case 'rerun': return RerunFormatter
+      case 'summary': return SummaryFormatter
+      default: throw new Error('Unknown formatter name "' + type + '".')
+    }
+  }
+
+  static getStepDefinitionSnippetBuilder({snippetInterface, snippetSyntax}) {
+    if (!snippetInterface) {
+      snippetInterface = 'callback'
+    }
+    let Syntax = JavascriptSnippetSyntax
+    if (snippetSyntax) {
+      const fullSyntaxPath = path.resolve(this.cwd, snippetSyntax);
+      Syntax = require(fullSyntaxPath);
+    }
+    const syntax = new Syntax(snippetInterface)
+    return new StepDefinitionSnippetBuilder(syntax)
   }
 }